requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
gitdb2==2.0.6
GitPython==3.0.3
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
sh==1.12.14
smmap2==2.0.5
sqlparse==0.3.0
urllib3==1.25.3
